Rename identifiers
