+++ macos/Runner/Configs/AppInfo.xcconfig
@@ -5,10 +5,10 @@
 // 'flutter create' template.
 
 // The application's name. By default this is also the title of the Flutter window.
-PRODUCT_NAME = roc_droid
+PRODUCT_NAME = roc-droid
 
 // The application's bundle identifier
-PRODUCT_BUNDLE_IDENTIFIER = com.example.rocDroid
+PRODUCT_BUNDLE_IDENTIFIER = org.rocstreaming.rocdroid
 
 // The copyright displayed in application information
-PRODUCT_COPYRIGHT = Copyright © 2024 com.example. All rights reserved.
+PRODUCT_COPYRIGHT = Copyright © Roc Streaming authors
