+++ pubspec.yaml
@@ -1,5 +1,5 @@
 name: roc_droid
-description: "A new Flutter project."
+description: "Roc for Android!"
 publish_to: 'none'
 version: 0.1.0
