+++ windows/CMakeLists.txt
@@ -4,7 +4,7 @@ project(roc_droid LANGUAGES CXX)
 
 # The name of the executable created for the application. Change this to change
 # the on-disk name of your application.
-set(BINARY_NAME "roc_droid")
+set(BINARY_NAME "roc-droid")
 
 # Explicitly opt in to modern CMake behaviors to avoid warnings with recent
 # versions of CMake.
